LayoutTests/svg/dynamic-updates/SVGCursorElement-svgdom-x-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x' property of the SVGCursorElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cursorElement.x.baseVal.value is 100
PASS cursorElement.x.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
